Replace all npm install/update/ci commands with a wrapper function that propagates errors
